# Optimize a scan of non state-chaning bytes with SSE2 instructions