--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "llparse",
-  "version": "7.1.1",
+  "version": "7.2.0",
   "description": "Compile incremental parsers to C code",
   "main": "lib/api.js",
   "types": "lib/api.d.ts",
@@ -34,7 +34,7 @@
   "devDependencies": {
     "@types/debug": "^4.1.5",
     "@types/mocha": "^8.0.3",
-    "@types/node": "^14.11.8",
+    "@types/node": "^14.14.14",
     "esm": "^3.2.25",
     "llparse-test-fixture": "^5.0.1",
     "mocha": "^9.2.2",
@@ -44,6 +44,8 @@
   },
   "dependencies": {
     "debug": "^4.2.0",
-    "llparse-frontend": "^3.0.0"
+    "llparse-frontend": "^3.0.0",
+    "z3-solver": "^4.12.2",
+    "synckit": "^0.8.5"
   }
 }