--- a/src/implementation/c/node/table-lookup.ts
+++ b/src/implementation/c/node/table-lookup.ts
@@ -1,7 +1,8 @@
 import * as assert from 'assert';
 import * as frontend from 'llparse-frontend';
+import { createSyncFn } from 'synckit'
 
-import { Compilation } from '../compilation';
+import { Compilation, SseFamily } from '../compilation';
 import { Node } from './base';
 
 const MAX_CHAR = 0xff;
@@ -65,7 +66,86 @@ export class TableLookup extends Node<frontend.node.TableLookup> {
     out.push('}');
   }
 
-  private buildSSE(out: string[]): boolean {
+  private buildSSE2(out: string[]): boolean {
+    // return false;
+    const ctx = this.compilation;
+
+    // Transformation is not supported atm
+    if (this.ref.transform && this.ref.transform.ref.name !== 'id') {
+      return false;
+    }
+
+    if (this.ref.edges.length !== 1) {
+      return false;
+    }
+
+    const edge = this.ref.edges[0];
+    if (edge.node.ref !== this.ref) {
+      return false;
+    }
+  
+    let initial_lut = new Array<number>(256).fill(0);
+    edge.keys.forEach(i => { initial_lut[i] = 1;});
+    
+    // the worker path must be absolute
+    const lutLowNibbleHighNibbleResolver = createSyncFn(require.resolve('./z3-lookup-solver'), {
+      tsRunner: 'ts-node', // optional, can be `'ts-node' | 'esbuild-register' | 'esbuild-runner' | 'tsx'`
+    })
+
+    // do whatever you want, you will get the result synchronously!
+    const result = lutLowNibbleHighNibbleResolver(initial_lut)
+
+    if (!result) {
+      return false;
+    }
+
+    const blob1 = ctx.blob(Buffer.from(result[0]), SSE_ALIGNMENT, SseFamily.SSE2);
+    const blob2 = ctx.blob(Buffer.from(result[1]), SSE_ALIGNMENT, SseFamily.SSE2);
+
+    out.push('#ifdef __SSE2__');
+    out.push(`if (${ctx.endPosArg()} - ${ctx.posArg()} >= 16) {`);
+    out.push('  int index;');
+    out.push('  __m128i lut_tlo;');
+    out.push('  __m128i lut_thi;');
+    out.push(`  lut_tlo = _mm_load_si128((__m128i const*) ${blob1});`);
+    out.push(`  lut_thi = _mm_load_si128((__m128i const*) ${blob2});`);
+    out.push('');
+    out.push(`  for( ;${ctx.endPosArg()} - ${ctx.posArg()} >= 16; ${ctx.posArg()} += 16) {`);
+
+    out.push('    __m128i lut_res_lo;');
+    out.push('    __m128i lut_res_hi;');
+    out.push('    __m128i input;');
+    out.push(`    input = _mm_loadu_si128((__m128i const*) ${ctx.posArg()});`);
+    out.push('');
+    out.push('    lut_res_lo = _mm_shuffle_epi8(lut_tlo, _mm_and_si128(input, _mm_set1_epi8(0x0F)));');
+    out.push('    lut_res_hi = _mm_shuffle_epi8(lut_thi, _mm_srli_epi16(_mm_and_si128(input, _mm_set1_epi8(0xF0)), 4));');
+    out.push('');
+    out.push('    input = _mm_cmpeq_epi8(_mm_and_si128(lut_res_lo, lut_res_hi), _mm_setzero_si128());');
+    out.push('    index = _mm_movemask_epi8(input);');
+    out.push('    if( 0 != index )');
+    out.push('    {');
+    out.push('      p += __builtin_ctz(index);');
+    {
+      const tmp: string[] = [];
+      this.tailTo(tmp, this.ref.otherwise!);
+      ctx.indent(out, tmp, '      ');
+    }
+    out.push('    }');
+
+    out.push('  }');
+    const tmp: string[] = [];
+    assert.strictEqual(edge.noAdvance, false);
+    this.tailTo(tmp, {
+      noAdvance: true,
+      node: edge.node,
+    });
+    ctx.indent(out, tmp, '  ');
+    out.push('}');
+    out.push('#endif /* __SSE2__ */');
+    return true;
+  }
+
+  private buildSSE42(out: string[]): boolean {
     const ctx = this.compilation;
 
     // Transformation is not supported atm
@@ -114,6 +194,7 @@ export class TableLookup extends Node<frontend.node.TableLookup> {
     }
 
     out.push('#ifdef __SSE4_2__');
+    out.push('// ${}');
     out.push(`if (${ctx.endPosArg()} - ${ctx.posArg()} >= 16) {`);
     out.push('  __m128i ranges;');
     out.push('  __m128i input;');
@@ -164,6 +245,13 @@ export class TableLookup extends Node<frontend.node.TableLookup> {
     out.push('#endif  /* __SSE4_2__ */');
 
     return true;
+  }
+
+  private buildSSE(out: string[]): boolean {
+    if (this.buildSSE2(out)){
+      return true;
+    }
+    return this.buildSSE42(out);
   }
 
   private buildTable(): ITable {