@ShogunPanda 2023-10-28
@anonrig @RafaelGSS I'm not familiar with SSE. Can you please chime in here?