@lemire 2023-11-07
@ShogunPanda I'll have a look when I am back in Montreal.