@ngrodzitski 2023-11-08
@lemire 
You are right, the source from which I learned this technique is simdjson. I've only added nice automation with z3 so that this works for generic LUT (if possible per se).

I can't say for Node.js, but I've recently ported [restinio](https://github.com/Stiffstream/restinio) which initially used old nodejs/http_parser to llhttp, thus I've given llhttp a closer look which eventually brings me here because my interest is that llhttp is fast (and correct of course).

So regarding a performance check I use [llhttp](https://github.com/nodejs/llhttp) repo which has a simple benchmark in it.

Here How I did it:
1. Cloned llparse and llhttp locally.
2. Do the changes in llparse.
3. Use [npm-link](https://docs.npmjs.com/cli/v10/commands/npm-link) so that my llocal llparse version is used in llhttp.
4. Rebuild llhttp and run a benchmark.

So my data (20% and 11%) is for llhttp.