@richardlau 2023-11-08
> > My question at this stage is whether any of this matters when built in Node.js because I don't see runtime dispatching anywhere (did I miss it?). So, in the current code, unless you compile with sse4.2 support, the fast path may not be called (ever).
> 
> That's correct. That would need to be done in the `deps/llhttp` first so we could, technically include Node.js, but I'm almost sure it would need a new dist creation (or at the very least - updating the current building flow) to have this support.
> 
> I assume this is a huge effort, right @richardlau?

I'm not entirely sure what is being asked. I'm not very familiar with llhttp and how it is generated using llparse -- for Node.js the gyp files used to build llhttp are:
- [`deps/llhttp/common.gypi`](https://github.com/nodejs/node/blob/main/deps/llhttp/common.gypi)
- [`deps/llhttp/llhttp.gyp`](https://github.com/nodejs/node/blob/main/deps/llhttp/llhttp.gyp)