@richardlau 2023-11-08
I don't think we'd need an arch specific build -- I think usually the dependency (in this case llhttp) would have to to detect and select the code to use at runtime. I don't know how easy that will be to do in llhttp so "huge effort" may or not be correct (albeit I'd expect most of the effort to be on the llhttp side rather than the Node.js side).
[sse4.2 in Node.js](https://github.com/search?q=repo%3Anodejs%2Fnode+sse4.2&type=code) shows hits in base64 and zlib (although https://github.com/nodejs/node/commit/ae115d68e010cf8860e15a5e8259e7dc520b26a5 commented this out for zlib).