@lemire 2023-11-08
@RafaelGSS Do you have a good http benchmark already that could be used if one wants to demonstrate benefits?